Fix bug introduced by warning patch.
Reverting the change in signedness of the `lane_id` variable fixes a
bug in optimized device code.

tl;dr:

It looks like the change in signedness of the `lane_id` introduced
a bug in a ReduceStep implementation, in which the expression

```
int offset = ...;
int last_lane = ...;
if (offset + lane_id <= last_lane) { /*... */ }
```

broke, because:

- `operator+(int, int)` returns `int`
- `operator+(int, uint32_t)` return `uint32_t`
- `last_lane` may be negative

My initial assumption that `last_lane` and `offset` would always
be positive was not correct.

In instances where `last_lane` is negative (which does happen
in the thrust.test.reduce TestReduceWithIndirectionDevice test),
this expression will always evaluate false, since an unsigned
value can't be <= a negative number.

So leave the ids as signed, since that seems necessary for this usecase.

+++ cub/warp/specializations/warp_reduce_shfl.cuh
@@ -98,10 +98,10 @@ struct WarpReduceShfl
     //---------------------------------------------------------------------
 
     /// Lane index in logical warp
-    uint32_t  lane_id;
+    int lane_id;
 
     /// Logical warp index in 32-thread physical warp
-    uint32_t warp_id;
+    int warp_id;
 
     /// 32-thread physical warp member mask of logical warp
     uint32_t member_mask;
@@ -115,7 +115,7 @@ struct WarpReduceShfl
     __device__ __forceinline__ WarpReduceShfl(
         TempStorage &/*temp_storage*/)
     {
-        lane_id = LaneId();
+        lane_id = static_cast<int>(LaneId());
         warp_id = 0;
         member_mask = 0xffffffffu >> (CUB_WARP_THREADS(PTX_ARCH) - LOGICAL_WARP_THREADS);
